keymap: allow to bind Esc key
However with this fast implementation Esc key will be in
handled state (in case if there is mapping with ALT+KEY)
so to bypass it to the vt the second Esc should be pressed.
But let left it as is, and fix it later, however I dont
this an issue with pressing Esc twice.

Signed-off-by: Vadym Kochan <vadim4j@gmail.com>